timer data change for redux
